Update to Code 2.0: Major change across all notebooks
This update represents a major change in the code. The code should now work with Pandas 2.0 and address other minor recent updates in code. It has nicely formatted chapter headers, the book is no longer forthcoming (it's here!), and it references the new Stack Downloader in the supplemental files, which makes downloading and cleaning the Stack Exchange archives very easy. If you have an earlier version it is recommended to update to this one in its entirety.